Public routing + controller logic
Public routing and controller logic for user pages, with the additions of the global footer and header, the layout dictates the overall structure of the website remains consistent throughout.
Note: the styling is extremely rudimentary, ensure that you revisit it at later stages when backend is complete and include the blade templates for the public pages even if they are immature implementations.